pyproject.toml
[project]
name = "crypto-news-aggregator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "flask>=3.1.2",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]
